pyproject.toml
[tool.poetry]
name = "seismos"
version = "0.1.0"
description = ""
authors = ["nixissmtp <gleb@simple2b.(none)>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.1"
Flask = "^2.0.1"
Flask-SQLAlchemy = "^2.5.1"
SQLAlchemy = "^1.4.20"
itsdangerous = "^2.0.1"
MarkupSafe = "^2.0.1"
python-dotenv = "^0.18.0"
Werkzeug = "^2.0.1"
Flask-RESTful = "^0.3.9"
mysqlclient = "^2.0.3"
SQLAlchemy-Utils = "^0.37.8"
Flask-JWT-Extended = "^4.2.3"
flasgger = "^0.9.5"
apispec = "^4.7.1"
marshmallow = "^3.12.2"
flask-apispec = "^0.11.0"
Flask-Migrate = "^3.0.1"
Flask-Cors = "^3.0.10"
lasio = "^0.29"
flake8 = "^3.9.2"
azure-storage-blob = "^12.9.0"
celery = "^5.2.1"
redis = "^4.0.1"
PyMySQL = "^1.0.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.6b0"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
